Knot images get deleted if type is changed from image to text

+++ src/components/index/KnotEditor.tsx
@@ -343,12 +343,11 @@ class KnotEditor extends React.Component<IProps, IState> {
     if (!knot || !data.value) {
       return;
     }
-
+    if ( knot.type === KnotType.IMAGE && data.value as KnotType === KnotType.TEXT  ) {
+      fetch(`${process.env.REACT_APP_API_BASE_PATH}/images/${knot.content}`, {method: "delete"});
+    }
     const { storyId, knotId } = this.props;
     knot.type = data.value as KnotType;
-    if ( knot.type === KnotType.TEXT  ) {
-        fetch(`${process.env.REACT_APP_API_BASE_PATH}/images/${knot.content}`, {method: "delete"});
-    }
     this.setState({
       loading: true,
     });
